requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
aiosmtplib==1.1.7
anyio==3.7.1
apispec==6.3.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
bech32==1.2.0
blinker==1.6.3
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cosmpy==0.9.1
ecdsa==0.18.0
Flask==3.0.0
frozenlist==1.4.0
googleapis-common-protos==1.60.0
grpcio==1.51.1
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
messages==0.8.0
msgpack==1.0.7
multidict==6.0.4
packaging==23.2
plyer==2.1.0
protobuf==4.24.4
pycryptodome==3.19.0
pydantic==1.10.13
python-dateutil==2.8.2
python-dotenv==1.0.0
referencing==0.30.2
requests==2.31.0
rfc3986==1.5.0
rpds-py==0.10.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
typing_extensions==4.8.0
uagents==0.6.2
urllib3==2.0.6
uvicorn==0.20.0
validus==0.3.0
websockets==10.4
Werkzeug==3.0.0
yarl==1.9.2